requirements.txt
# 01. Huggingface Download : pip install git+https://github.com/huggingface/transformers
# 02. Pytorch Download     : https://pytorch.org/get-started/locally/

sentencepiece==0.1.99
torchtext==0.15.2

# Tokenizer 
spacy==3.5.3
nltk==3.8.1